requirements.txt
streamlit==1.28.0 
langchain==0.1.0 
langchain-groq==0.1.0 
google-search-results==2.4.2 
python-dotenv==1.0.0 
duckduckgo-search==3.9.11 
faiss-cpu==1.7.4 
requests==2.31.0 
beautifulsoup4==4.12.0 
